requirements.txt
altgraph==0.17
asn1crypto==1.4.0
attrs==21.2.0
Automat==20.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorlog==3.1.0
constantly==15.1.0
cryptography==2.3
daemonize==2.4.7
httplib2==0.19.1
hyperlink==21.0.0
idna==2.6
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mock==2.0.0
ntplib==0.3.4
packaging==21.0
pbr==5.6.0
Pillow==8.3.1
plyvel==1.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.10.1
PyHamcrest==2.0.2
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pyOpenSSL==17.5.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PySocks==1.7.1
pytz==2021.1
qrcode==7.3
requests==2.26.0
rsa==4.7.2
service-identity==17.0.0
simplejson==3.11.1
six==1.13.0
Twisted==20.3.0
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
zope.interface==5.4.0

pyqrllib==1.2.4
xmssmt-wrapper==0.1.15
qsc-node-qscf==1.0.0

